refactor(snapshot): keep capability governance internal
