+++ src/snapshot-quality/verdict.ts
@@ -1,7 +1,5 @@
-import {
-  SNAPSHOT_BACKEND_CAPABILITIES,
-  type SnapshotQualityVerdict,
-} from '@agent-device/kernel/snapshot';
+import type { SnapshotQualityVerdict } from '@agent-device/kernel/snapshot';
+import { SNAPSHOT_BACKEND_CAPABILITIES } from './backend-capabilities.ts';
 
 const SNAPSHOT_QUALITY_STATES = new Set<SnapshotQualityVerdict['state']>([
   'healthy',
